# packages/typescript/api-types/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Roborean API",
    "version": "0.4.0"
  },
  "paths": {
...
    "/v1/projects": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "List Projects",
        "description": "List project summaries.\n\nArgs:\n    state: Shared repositories and run service.\n    _principal: Resolved caller identity (auth stub).\n\nReturns:\n    Summary rows for each stored project.",
        "operationId": "list_projects_v1_projects_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProjectSummary"
                  },
                  "type": "array",
                  "title": "Response List Projects V1 Projects Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Create Project",
        "description": "Create a stored project.\n\nArgs:\n    body: Create request containing the project document.\n    state: Shared repositories and run service.\n    _principal: Resolved caller identity (auth stub).\n\nReturns:\n    Redacted project detail for the created project.",
        "operationId": "create_project_v1_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/projects/{project_id}": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Get Project",
        "description": "Fetch one project.\n\nArgs:\n    project_id: Identifier of the project to load.\n    state: Shared repositories and run service.\n    _principal: Resolved caller identity (auth stub).\n\nReturns:\n    Redacted project detail for the requested project.",
        "operationId": "get_project_v1_projects__project_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "projects"
        ],
        "summary": "Update Project",
        "description": "Replace a project.\n\nArgs:\n    project_id: Identifier from the URL path.\n    body: Update request containing the replacement project.\n    state: Shared repositories and run service.\n    _principal: Resolved caller identity (auth stub).\n\nReturns:\n    Redacted project detail for the updated project.\n\nRaises:\n    ApiError: When the body project id does not match the URL.",
        "operationId": "update_project_v1_projects__project_id__put",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "projects"
        ],
        "summary": "Delete Project",
        "description": "Delete a project.\n\nArgs:\n    project_id: Identifier of the project to delete.\n    state: Shared repositories and run service.\n    _principal: Resolved caller identity (auth stub).\n\nReturns:\n    Empty response with HTTP 204 status.",
        "operationId": "delete_project_v1_projects__project_id__delete",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/projects/{project_id}/templates/{template_id}/content": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Get Template Content",
        "description": "Fetch one template file for a stored project.\n\nArgs:\n    project_id: Identifier of the project to load.\n    template_id: Template identifier from the project table.\n    state: Shared repositories and run service.\n    _principal: Resolved caller identity (auth stub).\n\nReturns:\n    Template file bytes and optional UTF-8 text body.",
        "operationId": "get_template_content_v1_projects__project_id__templates__template_id__content_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Template Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateContentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "projects"
        ],
        "summary": "Put Template Content",
        "description": "Replace one template file for a stored project.\n\nArgs:\n    project_id: Identifier of the project to update.\n    template_id: Template identifier from the project table.\n    body: Template bytes or UTF-8 text to persist.\n    state: Shared repositories and run service.\n    _principal: Resolved caller identity (auth stub).\n\nReturns:\n    Stored template file metadata.",
        "operationId": "put_template_content_v1_projects__project_id__templates__template_id__content_put",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Template Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateContentUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateContentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "projects"
        ],
        "summary": "Delete Template Content",
        "description": "Delete one template file from project storage.\n\nArgs:\n    project_id: Identifier of the project to update.\n    template_id: Template identifier from the project table.\n    state: Shared repositories and run service.\n    _principal: Resolved caller identity (auth stub).\n\nReturns:\n    Empty response with HTTP 204 status.",
        "operationId": "delete_template_content_v1_projects__project_id__templates__template_id__content_delete",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Template Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Bit": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "when": {
            "anyOf": [
              {
                "type": "boolean",
                "const": true
              },
              {
                "$ref": "#/components/schemas/RuleAst"
              }
            ],
            "title": "When"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          },
          "reads": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reads"
          },
          "writes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Writes"
          },
          "emits": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Emits"
          },
          "effectClass": {
            "$ref": "#/components/schemas/EffectClass"
          },
          "onError": {
            "$ref": "#/components/schemas/OnError"
          },
          "capabilities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Capabilities"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "type",
          "when",
          "config",
          "reads",
          "writes",
          "emits",
          "effectClass",
          "onError",
          "capabilities"
        ],
        "title": "Bit",
        "description": "An ordered conditional unit of project work.\n\nAttributes:\n    id: Unique identifier of the bit within the project.\n    type: The bit type implementation to execute.\n    label: Optional human-readable label for the bit.\n    when: Activation condition; either always-true or a rule\n        expression evaluated against the workspace.\n    config: Bit-specific configuration payload.\n    reads: Workspace keys the bit declares it reads.\n    writes: Workspace keys the bit declares it writes.\n    emits: Document operation kinds the bit declares it emits.\n    effect_class: Effect classification (``effectClass`` alias) used\n        for retry and idempotency handling.\n    on_error: Behavior to apply on failure (``onError`` alias).\n    capabilities: Capability identifiers required to run the bit."
      },
...
      "Exposure": {
        "type": "string",
        "enum": [
          "backendOnly",
          "redactedToClient",
          "clientVisible"
        ],
        "title": "Exposure",
        "description": "Controls whether a workspace value can reach clients.\n\nAttributes:\n    BACKEND_ONLY: The value never leaves the backend.\n    REDACTED_TO_CLIENT: The value reaches the client only after\n        redaction.\n    CLIENT_VISIBLE: The value may be sent to the client as-is."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Project": {
        "properties": {
          "schemaVersion": {
            "type": "string",
            "enum": [
              "1.0.0",
              "1.1.0"
            ],
            "title": "Schemaversion"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "pluginRequirements": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "type": "array",
            "title": "Pluginrequirements"
          },
          "workspace": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/Variable"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Workspace"
          },
          "bits": {
            "items": {
              "$ref": "#/components/schemas/Bit"
            },
            "type": "array",
            "title": "Bits"
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/DocumentDefinition"
            },
            "type": "array",
            "title": "Documents"
          },
          "templates": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "type": "array",
            "title": "Templates"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "schemaVersion",
          "id",
          "name",
          "pluginRequirements",
          "workspace",
          "bits",
          "documents",
          "templates",
          "metadata"
        ],
        "title": "Project",
        "description": "A portable, versioned Roborean project.\n\nAttributes:\n    schema_version: Schema version the project conforms to\n        (``schemaVersion`` alias): one of ``1.0.0`` or ``1.1.0``.\n    id: Unique identifier of the project.\n    name: Human-readable project name.\n    description: Optional human-readable project description.\n    plugin_requirements: Plugin dependency declarations\n        (``pluginRequirements`` alias).\n    workspace: Declared workspace variables, keyed by group name.\n    bits: Ordered list of bits that make up the project.\n    documents: Document definitions produced by the project.\n    templates: Template references used by the project's documents.\n    metadata: Free-form project metadata.\n    variables: Declared workspace variables, read from the\n        ``variables`` group of `workspace`."
      },
      "ProjectCreate": {
        "properties": {
          "project": {
            "$ref": "#/components/schemas/Project"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project"
        ],
        "title": "ProjectCreate",
        "description": "Create one stored project.\n\nAttributes:\n    project: Full project document to validate and persist."
      },
      "ProjectDetail": {
        "properties": {
          "project": {
            "additionalProperties": true,
            "type": "object",
            "title": "Project"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project"
        ],
        "title": "ProjectDetail",
        "description": "Redacted project payload.\n\nAttributes:\n    project: JSON-serializable project with secrets redacted."
      },
      "ProjectSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "schemaVersion": {
            "type": "string",
            "title": "Schemaversion"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "schemaVersion"
        ],
        "title": "ProjectSummary",
        "description": "List row for projects.\n\nAttributes:\n    id: Stable project identifier.\n    name: Human-readable project name.\n    schema_version: Project schema version (``schemaVersion``)."
      },
      "ProjectUpdate": {
        "properties": {
          "project": {
            "$ref": "#/components/schemas/Project"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project"
        ],
        "title": "ProjectUpdate",
        "description": "Replace a stored project.\n\nAttributes:\n    project: Replacement project document (id must match the URL)."
      },
      "PublicLiteral": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "public_literal",
            "title": "Kind"
          },
          "dataType": {
            "type": "string",
            "enum": [
              "string",
              "number",
              "boolean",
              "date"
            ],
            "title": "Datatype"
          },
          "value": {
            "title": "Value"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "dataType",
          "value"
        ],
        "title": "PublicLiteral",
        "description": "A public literal workspace value.\n\nAttributes:\n    kind: Discriminator identifying this value as a public literal.\n    data_type: The literal's declared type (``dataType`` alias): one\n        of ``string``, ``number``, ``boolean``, or ``date``.\n    value: The literal value itself."
      },
...
      "SecretRefValue": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "secret_ref",
            "title": "Kind"
          },
          "ref": {
            "type": "string",
            "title": "Ref"
          },
          "displayHint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Displayhint"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "ref"
        ],
        "title": "SecretRefValue",
        "description": "A reference to a secret held outside the portable project.\n\nAttributes:\n    kind: Discriminator identifying this value as a secret\n        reference.\n    ref: Opaque identifier used to resolve the secret at run time.\n    display_hint: Optional human-readable hint (``displayHint``\n        alias) shown in place of the secret value."
      },
...
      "TemplateContentResponse": {
        "properties": {
          "templateId": {
            "type": "string",
            "title": "Templateid"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "contentBase64": {
            "type": "string",
            "title": "Contentbase64"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "templateId",
          "path",
          "contentBase64"
        ],
        "title": "TemplateContentResponse",
        "description": "Template file payload returned by the API.\n\nAttributes:\n    template_id: Template identifier from the project table\n        (``templateId`` alias).\n    path: Relative package path for the template file.\n    content_base64: Base64-encoded template bytes\n        (``contentBase64`` alias).\n    text: UTF-8 decoded body when the file is textual."
      },
      "TemplateContentUpdate": {
        "properties": {
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "contentBase64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contentbase64"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "TemplateContentUpdate",
        "description": "Replace one template file.\n\nAttributes:\n    text: UTF-8 template body for text-like templates.\n    content_base64: Base64-encoded bytes for binary templates\n        (``contentBase64`` alias)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "Variable": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "schema": {
            "additionalProperties": true,
            "type": "object",
            "title": "Schema"
          },
          "defaultValue": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/PublicLiteral"
              },
              {
                "$ref": "#/components/schemas/SecretRefValue"
              },
              {
                "$ref": "#/components/schemas/EqToken"
              },
              {
                "$ref": "#/components/schemas/ShapeToken"
              },
              {
                "$ref": "#/components/schemas/Bucket"
              },
              {
                "$ref": "#/components/schemas/Redacted"
              }
            ],
            "title": "Defaultvalue",
            "discriminator": {
              "propertyName": "kind",
              "mapping": {
                "bucket": "#/components/schemas/Bucket",
                "eq_token": "#/components/schemas/EqToken",
                "public_literal": "#/components/schemas/PublicLiteral",
                "redacted": "#/components/schemas/Redacted",
                "secret_ref": "#/components/schemas/SecretRefValue",
                "shape_token": "#/components/schemas/ShapeToken"
              }
            }
          },
          "const": {
            "type": "boolean",
            "title": "Const",
            "default": false
          },
          "exposure": {
            "$ref": "#/components/schemas/Exposure"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "key",
          "schema",
          "defaultValue",
          "exposure"
        ],
        "title": "Variable",
        "description": "A declared workspace variable.\n\nAttributes:\n    key: The variable's unique workspace key.\n    schema_: JSON Schema (``schema`` alias) describing valid values.\n    default_value: Default workspace value (``defaultValue`` alias)\n        used when no override is provided.\n    const: Whether the variable's value is immutable once set.\n    exposure: Client-visibility policy for the variable's value.\n    description: Optional human-readable description of the\n        variable."
      }
    }
  }
}